// src/models/ProductModel.tsx
export default interface ProductModel {
    id: number;
    srno: number;
    activityRoleName: string;
    serviceName: string;
    categoryName: string;
    productCode: string;
    productName: string;
    hsnSacCode: string;
    gstRate: number;
    unitOfSales: string;
    display: string;
    action: string;
  }
  